Renaming modules' name to snake case.
